* master: (108 commits)
  Add very simple search form tests
  MBS-3292, Remove id from cloned buttons in MB.utility.setDefaultAction.
  MBS-1543, Attribute BBC supplied icons.
  MBS-2817, trim submitted label names before searching for them on missing entities page.
  Fix test failures found by new Edit::Medium::Edit tests
  Add tests for Edit::Medium::Edit
  MBS-3224, fix crash in Release Editor by including a page id as hidden input on all pages.
  MBS-3261: Allow users to delete their own accounts
  MBS-3306: Use the is_account_admin predicate when deciding whether to show the delete/adjust flags tabs
  MBS-3113: Only allow entering URLs that use HTTP, HTTPs or FTP protocols
  MBS-3301: When editing mediums, do allow some new recordings to be created
  MBS-2829: Use murdos' /ws/2/discid algorithm
  Don't try and auto-approve closed edits in the release editor
  Add missing delete_user page
  MBS-2955: Fix the display of relationship attributes in edit relationship type edits
  MBS-3012: Cast ratings to integers when submitting ratings
  MBS-3141: Re-name edit search input elements after initializing the autocompleter
  MBS-3247: Use descriptive_link to link to recordings in remove ISRC edit
  MBS-3253: Select distinct edits in edit searches
  MBS-2881: Introduce URL->affiliate_url for linking to affiliate sites
  ...

Conflicts:
	root/static/scripts/common/MB/edit_search.js

+++ lib/MusicBrainz/Server/Data/Collection.pm
@@ -209,6 +209,7 @@ sub update
 sub delete
 {
     my ($self, @collection_ids) = @_;
+    return unless @collection_ids;
 
     $self->sql->auto_commit;
     $self->sql->do('DELETE FROM editor_collection_release
@@ -217,6 +218,16 @@ sub delete
     $self->sql->do('DELETE FROM editor_collection
                     WHERE id IN (' . placeholders(@collection_ids) . ')', @collection_ids);
     return;
+}
+
+sub delete_editor {
+    my ($self, $editor_id) = @_;
+    $self->delete(
+        @{ $self->sql->select_single_column_array(
+            'SELECT id FROM editor_collection WHERE editor = ?',
+            $editor_id
+        ) }
+    );
 }
 
 sub _hash_to_row
